feat(docs): merge examples scripts in one
