+++ scripts/gulp/tasks/config.ts
@@ -2,3 +2,4 @@ export const BUILD_DIR = './.ng_build';
 export const LIB_DIR = './src/.lib';
 export const PLAYGROUND_ROOT = './src/playground/';
 export const DOCS_OUTPUT = './docs/output.json';
+export const EXTENSIONS = ['ts', 'html', 'scss'];
